typed play tournament button
